requirements.txt
asgiref==3.6.0
Brotli==1.0.9
dj-database-url==2.0.0
Django==4.2
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.0
tzdata==2023.3
whitenoise==6.4.0
